refactor: move window overlay support
